consolidate storybook files into one directroy
