+++ styleguide/config.js
@@ -1,7 +1,7 @@
 import { configure } from '@storybook/react'
 
 const loadStories = () => {
-  require('../stories/index.js')
+  require('./stories/index.js')
 }
 
 configure(loadStories, module)
