+++ styleguide/stories/typography/index.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import { muiTheme } from 'storybook-addon-material-ui'
 import { createMuiTheme } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
-import theme from '../../gui/src/theme'
+import theme from '../../../gui/src/theme'
 
 const customTheme = createMuiTheme(theme)
